ea8300: Switch to DSA configuration
Changes to EA8300 network init and preinit for DSA

Device's primary MAC address (from devinfo partition)
assigned to lan, secondary assigned to  wan

Preinit is on lan1. This single-port behavior is consistent with
comments in package/base-files/files/lib/preinit/10_indicate_preinit
See 467cd37 for where support of DSA was added to 10_indicate_preinit

At this commit, eth0 is still getting the common-to-all MAC address
from the U-Boot environment, ethaddr=00:03:7f:ba:db:ad
As this appears to be the "base" for the VAPs, these MACs and BSSIDs
may be duplicated across different instances of this family.

Signed-off-by: Jeff Kletsky <git-commits@allycomm.com>