ipq40xx: ea8300: mr8300: Convert to DSA driver
NB: Dependent on acceptance of
    https://github.com/openwrt/openwrt/pull/4721
    or equivalent

Adapt these related Linksys devices to the DSA framework

Device's primary MAC address (from devinfo partition)
assigned to lan, secondary assigned to wan

Preinit is on lan1. This port is labeled "1" under "Ethernet".

This single-port preinit behavior is consistent with comments in
package/base-files/files/lib/preinit/10_indicate_preinit
See 467cd37 for where support of DSA was added to 10_indicate_preinit

The Linksys EA8300 has ethaddr=00:03:7f:ba:db:ad in the U-Boot env,
apparently for every instance. When eth0 is configured, it picks
this up. With DSA, it is lan1 that is reconfigured, not eth0.
It appears that the VAPs' BSSIDs are based off eth0. As a result,
both eth0 and the BSSIDs would not be unique by instance.

To prevent this, the eth0 MAC is based off the "factory configured"
one in the devinfo partition, by setting the locally adminstered bit.

Signed-off-by: Jeff Kletsky <git-commits@allycomm.com>