ea8300: DSA: Set eth0 MAC locally administered
The Linksys EA8300 has ethaddr=00:03:7f:ba:db:ad in the U-Boot env,
apparently for every instance. When eth0 is configured, it picks
this up. With DSA, it is lan1 that is reconfigured, not eth0.
It appears that the VAPs' BSSIDs are based off eth0. As a result,
prior to this patch, the BSSIDs would not be unique by instance.

This patch overrides the eth0 MAC with a MAC address based off
the "factory configured" one in the devinfo partition,
by setting the locally adminstered bit.

This patch has been split from the main DSA patch for the EA8300
so that the reasoning is documented through `git blame`
and, should later changes allow, comfortably reverted.

Signed-off-by: Jeff Kletsky <git-commits@allycomm.com>