ipq40xx: ea8300: mr8300: DSA: Adjust MAC addresses
With more complex configurations and the previous, simple assignments,
MAC collisions can occur between Ethernet ports as well as with WiFi.

This patch assigns the "label MAC" to "wan" and the next sequential
to "lan1". lan2-4 are assigned locally administered MAC addresses,
sequential from lan1 with its LA bit set. "eth0" gets assigned
the next in sequence, which appears to be the "base" for the WiFi MACs.

label_mac is also defined and now available in board.json

NB: Will be merged with previous commit prior to PR

Signed-off-by: Jeff Kletsky <git-commits@allycomm.com>